+++ src/Application/DependencyInjection.cs
@@ -3,7 +3,6 @@ using Microsoft.Extensions.Options;
 using NotionMarkdownConverter.Application.Abstractions;
 using NotionMarkdownConverter.Application.Configuration;
 using NotionMarkdownConverter.Application.Services;
-using NotionMarkdownConverter.Domain.Mappers;
 using NotionMarkdownConverter.Domain.Markdown.Converters;
 using NotionMarkdownConverter.Domain.Transformers;
 using NotionMarkdownConverter.Domain.Transformers.Strategies;
