+++ tests/Unit/PageExportEligibilityCheckerTests.cs
@@ -1,4 +1,4 @@
-using NotionMarkdownConverter.Application.Services;
+using NotionMarkdownConverter.Extract;
 using NotionMarkdownConverter.Shared.Enums;
 using NotionMarkdownConverter.Shared.Models;
